requirements/local.txt
-r production.txt
coverage
django-nose
django-test-without-migrations
flake8
ipdb
mixer
